fixup: add more gauth setup details
* And move logic to its own file.
